* commit 'couchbase/mad-hatter':
  MB-36923: Allow "default" and "disk_io_optimized" for ThreadCount

+++ daemon/settings.h
@@ -801,11 +801,11 @@ public:
         notify_changed("parent_identifier");
     }
 
-    size_t getNumReaderThreads() const {
+    int getNumReaderThreads() const {
         return num_reader_threads.load(std::memory_order_acquire);
     }
 
-    void setNumReaderThreads(size_t val) {
+    void setNumReaderThreads(int val) {
         num_reader_threads.store(val, std::memory_order_release);
         has.num_reader_threads = true;
         notify_changed("num_reader_threads");
@@ -1012,8 +1012,8 @@ protected:
 
     void notify_changed(const std::string& key);
 
-    std::atomic<size_t> num_reader_threads{0};
-    std::atomic<size_t> num_writer_threads{0};
+    std::atomic<int> num_reader_threads{0};
+    std::atomic<int> num_writer_threads{0};
 
 public:
     /**
